third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/contentWrappers-expected.txt

Test HTMLTemplateElement content node wrappers have the correct lifetime.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS span.tagName is "SPAN"
PASS span.expando is "present"
PASS span.tagName is "SPAN"
PASS span.expando is "present"
PASS successfullyParsed is true

TEST COMPLETE
